requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aioconsole==0.4.1
    # via -r requirements.in
aiodns==3.0.0
    # via aiohttp
aiofiles==0.8.0
    # via -r requirements.in
aiohttp[speedups]==3.8.1
    # via
    #   -r requirements.in
    #   aiohttp-jinja2
    #   aioresponses
    #   pytest-aiohttp
aiohttp-jinja2==1.5
    # via -r requirements.in
aioresponses==0.7.3
    # via -r requirements.in
aiosignal==1.2.0
    # via aiohttp
anyio==3.5.0
    # via watchfiles
astroid==2.11.3
    # via pylint
async-timeout==4.0.2
    # via aiohttp
asyncpg==0.25.0
    # via -r requirements.in
attrs==21.4.0
    # via
    #   aiohttp
    #   pytest
    #   pytest-mypy
backoff==1.11.1
    # via opentelemetry-exporter-otlp-proto-grpc
beautifulsoup4==4.11.1
    # via -r requirements.in
black==22.3.0
    # via
    #   -r requirements.in
    #   pytest-black
brotli==1.0.9
    # via aiohttp
cchardet==2.1.7
    # via aiohttp
cffi==1.15.0
    # via pycares
charset-normalizer==2.0.12
    # via aiohttp
click==8.1.3
    # via
    #   black
    #   pip-tools
coverage[toml]==6.3.2
    # via pytest-cov
deprecated==1.2.13
    # via opentelemetry-api
dill==0.3.4
    # via pylint
execnet==1.9.0
    # via pytest-xdist
filelock==3.6.0
    # via pytest-mypy
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
googleapis-common-protos==1.56.0
    # via opentelemetry-exporter-otlp-proto-grpc
grpcio==1.44.0
    # via opentelemetry-exporter-otlp-proto-grpc
humanize==4.0.0
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   yarl
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via
    #   -r requirements.in
    #   pylint
    #   pytest-isort
jinja2==3.1.2
    # via aiohttp-jinja2
lazy-object-proxy==1.7.1
    # via astroid
lxml==4.8.0
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
mccabe==0.7.0
    # via pylint
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
mypy==0.950
    # via
    #   -r requirements.in
    #   pytest-mypy
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
opentelemetry-api==1.11.1
    # via
    #   -r requirements.in
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asyncpg
    #   opentelemetry-instrumentation-jinja2
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-sdk
    #   pytest-opentelemetry
opentelemetry-distro==0.30b1
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-grpc==1.11.1
    # via
    #   -r requirements.in
    #   pytest-opentelemetry
opentelemetry-instrumentation==0.30b1
    # via
    #   opentelemetry-distro
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asyncpg
    #   opentelemetry-instrumentation-jinja2
    #   opentelemetry-instrumentation-logging
opentelemetry-instrumentation-aiohttp-client==0.30b1
    # via -r requirements.in
opentelemetry-instrumentation-asyncpg==0.30b1
    # via -r requirements.in
opentelemetry-instrumentation-jinja2==0.30b1
    # via -r requirements.in
opentelemetry-instrumentation-logging==0.30b1
    # via -r requirements.in
opentelemetry-proto==1.11.1
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.11.1
    # via
    #   -r requirements.in
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   pytest-opentelemetry
opentelemetry-semantic-conventions==0.30b1
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asyncpg
    #   opentelemetry-sdk
    #   pytest-opentelemetry
opentelemetry-util-http==0.30b1
    # via opentelemetry-instrumentation-aiohttp-client
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.0
    # via -r requirements.in
platformdirs==2.5.2
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
protobuf==3.20.1
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
psutil==5.9.0
    # via pytest-opentelemetry
py==1.11.0
    # via
    #   pytest
    #   pytest-forked
pycares==4.1.2
    # via aiodns
pycparser==2.21
    # via cffi
pylint==2.13.7
    # via
    #   -r requirements.in
    #   pytest-pylint
pyparsing==3.0.8
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements.in
    #   pytest-aiohttp
    #   pytest-asyncio
    #   pytest-black
    #   pytest-cov
    #   pytest-forked
    #   pytest-isort
    #   pytest-mypy
    #   pytest-opentelemetry
    #   pytest-pylint
    #   pytest-xdist
pytest-aiohttp==1.0.4
    # via -r requirements.in
pytest-asyncio==0.18.3
    # via
    #   -r requirements.in
    #   pytest-aiohttp
pytest-black==0.3.12
    # via -r requirements.in
pytest-cov==3.0.0
    # via -r requirements.in
pytest-forked==1.4.0
    # via pytest-xdist
pytest-isort==3.0.0
    # via -r requirements.in
pytest-mypy==0.9.1
    # via -r requirements.in
pytest-opentelemetry==0.2.1
    # via -r requirements.in
pytest-pylint==0.18.0
    # via -r requirements.in
pytest-xdist==2.5.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
six==1.16.0
    # via
    #   grpcio
    #   python-dateutil
sniffio==1.2.0
    # via anyio
soupsieve==2.3.2.post1
    # via beautifulsoup4
toml==0.10.2
    # via
    #   pytest-black
    #   pytest-pylint
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   mypy
    #   pep517
    #   pylint
    #   pytest
types-aiofiles==0.8.8
    # via -r requirements.in
types-beautifulsoup4==4.10.20
    # via -r requirements.in
types-python-dateutil==2.8.14
    # via -r requirements.in
typing-extensions==4.2.0
    # via
    #   astroid
    #   black
    #   mypy
    #   opentelemetry-sdk
    #   pylint
watchfiles==0.13
    # via -r requirements.in
wheel==0.37.1
    # via pip-tools
wrapt==1.14.0
    # via
    #   astroid
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-jinja2
yarl==1.7.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
